HandleImage to component
