+++ src/actions/FilmActions.jsx
@@ -1,7 +1,7 @@
-import Parameters from "@/api/Parameters";
-import Soliciter from "@/api/Soliciter";
-import db from "@/api/db";
-import RequestMode from "@/api/RequestMode";
+import Parameters from "@/services/Parameters";
+import Soliciter from "@/services/Soliciter";
+import db from "@/services/db";
+import RequestMode from "@/services/RequestMode";
 import ACTIONS from "@/utils/ActionsCreators/FilmTypes";
 import useDispatch from "@/hooks/useDispatch";
 import useStore from "@/hooks/useStore";
